* Tried using makeWrapResolversPlugin to integrate the graphql-live-query-patch library, but seems it's not low-level enough. (the object returned by the func in LQHelper.ts keeps being interpreted as merely the "shape" of the data -- the graphql code then tries to resolve the object into subfields...)
My guess is that the makeWrapResolversPlugin can only do the "return whatever it wants" behavior, if it's used for a single field; when used for a root query/mutation, its result is always "resolved" as subfields. Thus, I need to use a lower-level system. (Next I'll try the hook approach described here: https://www.graphile.org/postgraphile/extending-raw/#wrapping-an-existing-resolver)